unity-copilot/python/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.12.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
faiss-cpu==1.11.0
fastapi==0.110.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.18.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.36.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.66.1
transformers==4.35.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.27.0
watchfiles==1.0.5
websockets==15.0.1
